python_modules/requirements.txt
requests==2.31.0
pillow==10.1.0
opencv-python==4.8.1.78
python-telegram-bot==20.7
python-dotenv==1.0.0